requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.95.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
packaging==23.2
pydantic==1.10.14
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
